support null-safety & 4.0 beta
