+++ dio/lib/src/entry/dio_for_browser.dart
@@ -1,3 +1,4 @@
+import '../dio_mixin.dart';
 import '../options.dart';
 import '../dio.dart';
 import '../adapters/browser_adapter.dart';
