+++ dio/lib/src/options.dart
@@ -41,7 +41,29 @@ enum ResponseType {
   bytes
 }
 
+/// CollectionFormat specifies the array format
+/// (a single parameter with multiple parameter or multiple parameters with the same name)
+/// and the separator for array items.
+enum CollectionFormat {
+  // Comma-separated values
+  // e.g. (foo,bar,baz)
+  csv,
+  // Space-separated values
+  // e.g. (foo bar baz)
+  ssv,
+  // Tab-separated values
+  // e.g. (foo\tbar\tbaz)
+  tsv,
+  // Pipe-separated values
+  // e.g. (foo|bar|baz)
+  pipes,
+  // Multiple parameter instances rather than multiple values.
+  // e.g. (foo=value&foo=another_value)
+  multi,
+}
+
 typedef ValidateStatus = bool Function(int? status);
+
 typedef ResponseDecoder = String Function(
     List<int> responseBytes, RequestOptions options, ResponseBody responseBody);
 typedef RequestEncoder = List<int> Function(
@@ -52,7 +74,7 @@ typedef RequestEncoder = List<int> Function(
 class BaseOptions extends _RequestConfig {
   BaseOptions({
     String? method,
-    this.connectTimeout,
+    int? connectTimeout,
     int? receiveTimeout,
     int? sendTimeout,
     this.baseUrl = '',
@@ -67,6 +89,7 @@ class BaseOptions extends _RequestConfig {
     int? maxRedirects,
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
+    CollectionFormat? collectionFormat,
   })  : queryParameters = queryParameters ?? {},
         super(
           method: method,
@@ -82,7 +105,10 @@ class BaseOptions extends _RequestConfig {
           maxRedirects: maxRedirects,
           requestEncoder: requestEncoder,
           responseDecoder: responseDecoder,
-        );
+          collectionFormat: collectionFormat,
+        ) {
+    this.connectTimeout = connectTimeout ?? 0;
+  }
 
   /// Create a Option from current instance with merging attributes.
   BaseOptions merge({
@@ -103,6 +129,7 @@ class BaseOptions extends _RequestConfig {
     int? maxRedirects,
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
+    CollectionFormat? collectionFormat,
   }) {
     return BaseOptions(
       method: method ?? this.method,
@@ -122,6 +149,7 @@ class BaseOptions extends _RequestConfig {
       maxRedirects: maxRedirects ?? this.maxRedirects,
       requestEncoder: requestEncoder ?? this.requestEncoder,
       responseDecoder: responseDecoder ?? this.responseDecoder,
+      collectionFormat: collectionFormat ?? this.collectionFormat,
     );
   }
 
@@ -134,40 +162,27 @@ class BaseOptions extends _RequestConfig {
   /// Timeout in milliseconds for opening url.
   /// [Dio] will throw the [DioError] with [DioErrorType.CONNECT_TIMEOUT] type
   ///  when time out.
-  int? connectTimeout;
+  late int connectTimeout;
 }
 
 /// Every request can pass an [Options] object which will be merged with [Dio.options]
-class Options extends _RequestConfig {
+class Options {
   Options({
-    String? method,
-    int? sendTimeout,
-    int? receiveTimeout,
-    Map<String, dynamic>? extra,
-    Map<String, dynamic>? headers,
-    ResponseType? responseType,
-    String? contentType,
-    ValidateStatus? validateStatus,
-    bool? receiveDataWhenStatusError,
-    bool? followRedirects,
-    int? maxRedirects,
-    RequestEncoder? requestEncoder,
-    ResponseDecoder? responseDecoder,
-  }) : super(
-          method: method,
-          sendTimeout: sendTimeout,
-          receiveTimeout: receiveTimeout,
-          extra: extra,
-          headers: headers,
-          responseType: responseType,
-          contentType: contentType,
-          validateStatus: validateStatus,
-          receiveDataWhenStatusError: receiveDataWhenStatusError,
-          followRedirects: followRedirects,
-          maxRedirects: maxRedirects,
-          requestEncoder: requestEncoder,
-          responseDecoder: responseDecoder,
-        );
+    this.method,
+    this.sendTimeout,
+    this.receiveTimeout,
+    this.extra,
+    this.headers,
+    this.responseType,
+    this.contentType,
+    this.validateStatus,
+    this.receiveDataWhenStatusError,
+    this.followRedirects,
+    this.maxRedirects,
+    this.requestEncoder,
+    this.responseDecoder,
+    this.collectionFormat,
+  }) ;
 
   /// Create a Option from current instance with merging attributes.
   Options merge({
@@ -184,13 +199,14 @@ class Options extends _RequestConfig {
     int? maxRedirects,
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
+    CollectionFormat? collectionFormat,
   }) {
     return Options(
       method: method ?? this.method,
       sendTimeout: sendTimeout ?? this.sendTimeout,
       receiveTimeout: receiveTimeout ?? this.receiveTimeout,
-      extra: extra ?? Map.from(this.extra),
-      headers: headers ?? Map.from(this.headers),
+      extra: extra ?? this.extra,
+      headers: headers ?? this.headers,
       responseType: responseType ?? this.responseType,
       contentType: contentType ?? this.contentType,
       validateStatus: validateStatus ?? this.validateStatus,
@@ -200,23 +216,154 @@ class Options extends _RequestConfig {
       maxRedirects: maxRedirects ?? this.maxRedirects,
       requestEncoder: requestEncoder ?? this.requestEncoder,
       responseDecoder: responseDecoder ?? this.responseDecoder,
+      collectionFormat: collectionFormat ?? this.collectionFormat,
     );
   }
+
+  RequestOptions compose(
+    BaseOptions baseOpt,
+    String path,
+    data, {
+    Map<String, dynamic>? queryParameters,
+    CancelToken? cancelToken,
+    Options? options,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) {
+    var query = <String, dynamic>{};
+    if (queryParameters != null) query.addAll(queryParameters);
+    query.addAll(baseOpt.queryParameters);
+
+    var _headers = Map<String,dynamic>.from(baseOpt.headers);
+    if(headers!=null){
+      _headers.addAll(headers!);
+    }
+
+    var _extra = Map<String,dynamic>.from(baseOpt.extra);
+    if(extra!=null){
+      _extra.addAll(extra!);
+    }
+
+    var requestOptions = RequestOptions(
+      method: (method ?? baseOpt.method).toUpperCase(),
+      headers: _headers,
+      extra: _extra,
+      baseUrl: baseOpt.baseUrl,
+      path: path,
+      data: data,
+      connectTimeout: baseOpt.connectTimeout,
+      sendTimeout: sendTimeout ?? baseOpt.sendTimeout,
+      receiveTimeout: receiveTimeout ?? baseOpt.receiveTimeout,
+      responseType: responseType ?? baseOpt.responseType,
+      contentType:
+          contentType ?? baseOpt.contentType ?? Headers.jsonContentType,
+      validateStatus: validateStatus ?? baseOpt.validateStatus,
+      receiveDataWhenStatusError:
+          receiveDataWhenStatusError ?? baseOpt.receiveDataWhenStatusError,
+      followRedirects: followRedirects ?? baseOpt.followRedirects,
+      maxRedirects: maxRedirects ?? baseOpt.maxRedirects,
+      queryParameters: query,
+      requestEncoder: requestEncoder ?? baseOpt.requestEncoder,
+      responseDecoder: responseDecoder ?? baseOpt.responseDecoder,
+    );
+
+    requestOptions.onReceiveProgress = onReceiveProgress;
+    requestOptions.onSendProgress = onSendProgress;
+    requestOptions.cancelToken = cancelToken;
+
+    return requestOptions;
+  }
+
+  /// Http method.
+  String? method;
+
+  /// Http request headers. The keys of initial headers will be converted to lowercase,
+  /// for example 'Content-Type' will be converted to 'content-type'.
+  ///
+  /// You should use lowercase as the key name when you need to set the request header.
+  Map<String, dynamic>? headers;
+
+  /// Timeout in milliseconds for sending data.
+  /// [Dio] will throw the [DioError] with [DioErrorType.SEND_TIMEOUT] type
+  ///  when time out.
+  int? sendTimeout;
+
+  ///  Timeout in milliseconds for receiving data.
+  ///  [Dio] will throw the [DioError] with [DioErrorType.RECEIVE_TIMEOUT] type
+  ///  when time out.
+  ///
+  /// [0] meanings no timeout limit.
+  int? receiveTimeout;
+
+  /// The request Content-Type. The default value is [ContentType.json].
+  /// If you want to encode request body with 'application/x-www-form-urlencoded',
+  /// you can set `ContentType.parse('application/x-www-form-urlencoded')`, and [Dio]
+  /// will automatically encode the request body.
+  String? contentType;
+
+  /// [responseType] indicates the type of data that the server will respond with
+  /// options which defined in [ResponseType] are `json`, `stream`, `plain`.
+  ///
+  /// The default value is `json`, dio will parse response string to json object automatically
+  /// when the content-type of response is 'application/json'.
+  ///
+  /// If you want to receive response data with binary bytes, for example,
+  /// downloading a image, use `stream`.
+  ///
+  /// If you want to receive the response data with String, use `plain`.
+  ///
+  /// If you want to receive the response data with  original bytes,
+  /// that's to say the type of [Response.data] will be List<int>, use `bytes`
+  ResponseType? responseType;
+
+  /// `validateStatus` defines whether the request is successful for a given
+  /// HTTP response status code. If `validateStatus` returns `true` ,
+  /// the request will be perceived as successful; otherwise, considered as failed.
+  ValidateStatus? validateStatus;
+
+  /// Whether receiving response data when http status code is not successful.
+  bool? receiveDataWhenStatusError;
+
+  /// Custom field that you can retrieve it later in [Interceptor]、[Transformer] and the [Response] object.
+  Map<String, dynamic>? extra;
+
+  /// see [HttpClientRequest.followRedirects]
+  bool? followRedirects;
+
+  /// Set this property to the maximum number of redirects to follow
+  /// when [followRedirects] is `true`. If this number is exceeded
+  /// an error event will be added with a [RedirectException].
+  ///
+  /// The default value is 5.
+  int? maxRedirects;
+
+  /// The default request encoder is utf8encoder, you can set custom
+  /// encoder by this option.
+  RequestEncoder? requestEncoder;
+
+  /// The default response decoder is utf8decoder, you can set custom
+  /// decoder by this option, it will be used in [Transformer].
+  ResponseDecoder? responseDecoder;
+
+  /// [collectionFormat] indicates the format of collection data in request
+  /// options which defined in [CollectionFormat] are `csv`, `ssv`, `tsv`, `pipes`, `multi`.
+  /// The default value is `csv`
+  CollectionFormat? collectionFormat;
 }
 
-class RequestOptions extends Options {
+class RequestOptions extends BaseOptions {
   RequestOptions({
     String? method,
     int? sendTimeout,
     int? receiveTimeout,
-    this.connectTimeout,
+    int? connectTimeout,
     this.data,
     this.path = '',
     Map<String, dynamic>? queryParameters,
-    this.baseUrl = '',
     this.onReceiveProgress,
     this.onSendProgress,
     this.cancelToken,
+    String? baseUrl,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
     ResponseType? responseType,
@@ -227,12 +374,15 @@ class RequestOptions extends Options {
     int? maxRedirects,
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
-  })  : queryParameters = queryParameters ?? {},
-        super(
+    CollectionFormat? collectionFormat,
+  }) : super(
           method: method,
           sendTimeout: sendTimeout,
           receiveTimeout: receiveTimeout,
+          connectTimeout: connectTimeout,
+          queryParameters: queryParameters,
           extra: extra,
+          baseUrl: baseUrl ?? '',
           headers: headers,
           responseType: responseType,
           contentType: contentType,
@@ -242,6 +392,7 @@ class RequestOptions extends Options {
           maxRedirects: maxRedirects,
           requestEncoder: requestEncoder,
           responseDecoder: responseDecoder,
+          collectionFormat: collectionFormat,
         );
 
   /// Create a Option from current instance with merging attributes.
@@ -268,31 +419,32 @@ class RequestOptions extends Options {
     int? maxRedirects,
     RequestEncoder? requestEncoder,
     ResponseDecoder? responseDecoder,
+    CollectionFormat? collectionFormat,
   }) {
     return RequestOptions(
-      method: method ?? this.method,
-      sendTimeout: sendTimeout ?? this.sendTimeout,
-      receiveTimeout: receiveTimeout ?? this.receiveTimeout,
-      connectTimeout: connectTimeout ?? this.connectTimeout,
-      data: data ?? this.data,
-      path: path ?? this.path,
-      queryParameters: queryParameters ?? this.queryParameters,
-      baseUrl: baseUrl ?? this.baseUrl,
-      onReceiveProgress: onReceiveProgress ?? this.onReceiveProgress,
-      onSendProgress: onSendProgress ?? this.onSendProgress,
-      cancelToken: cancelToken ?? this.cancelToken,
-      extra: extra ?? Map.from(this.extra),
-      headers: headers ?? Map.from(this.headers),
-      responseType: responseType ?? this.responseType,
-      contentType: contentType ?? this.contentType,
-      validateStatus: validateStatus ?? this.validateStatus,
-      receiveDataWhenStatusError:
-          receiveDataWhenStatusError ?? this.receiveDataWhenStatusError,
-      followRedirects: followRedirects ?? this.followRedirects,
-      maxRedirects: maxRedirects ?? this.maxRedirects,
-      requestEncoder: requestEncoder ?? this.requestEncoder,
-      responseDecoder: responseDecoder ?? this.responseDecoder,
-    );
+        method: method ?? this.method,
+        sendTimeout: sendTimeout ?? this.sendTimeout,
+        receiveTimeout: receiveTimeout ?? this.receiveTimeout,
+        connectTimeout: connectTimeout ?? this.connectTimeout,
+        data: data ?? this.data,
+        path: path ?? this.path,
+        queryParameters: queryParameters ?? this.queryParameters,
+        baseUrl: baseUrl ?? this.baseUrl,
+        onReceiveProgress: onReceiveProgress ?? this.onReceiveProgress,
+        onSendProgress: onSendProgress ?? this.onSendProgress,
+        cancelToken: cancelToken ?? this.cancelToken,
+        extra: extra ?? Map.from(this.extra),
+        headers: headers ?? Map.from(this.headers),
+        responseType: responseType ?? this.responseType,
+        contentType: contentType ?? this.contentType,
+        validateStatus: validateStatus ?? this.validateStatus,
+        receiveDataWhenStatusError:
+            receiveDataWhenStatusError ?? this.receiveDataWhenStatusError,
+        followRedirects: followRedirects ?? this.followRedirects,
+        maxRedirects: maxRedirects ?? this.maxRedirects,
+        requestEncoder: requestEncoder ?? this.requestEncoder,
+        responseDecoder: responseDecoder ?? this.responseDecoder,
+        collectionFormat: collectionFormat ?? this.collectionFormat);
   }
 
   /// generate uri
@@ -303,7 +455,7 @@ class RequestOptions extends Options {
       var s = _url.split(':/');
       _url = s[0] + ':/' + s[1].replaceAll('//', '/');
     }
-    var query = Transformer.urlEncodeMap(queryParameters);
+    var query = Transformer.urlEncodeMap(queryParameters, collectionFormat);
     if (query.isNotEmpty) {
       _url += (_url.contains('?') ? '&' : '?') + query;
     }
@@ -312,69 +464,75 @@ class RequestOptions extends Options {
   }
 
   /// Request data, can be any type.
-  dynamic data;
-
-  /// Request base url, it can contain sub path, like: 'https://www.google.com/api/'.
-  late String baseUrl;
+  dynamic? data;
 
   /// If the `path` starts with 'http(s)', the `baseURL` will be ignored, otherwise,
   /// it will be combined and then resolved with the baseUrl.
   String path = '';
-
-  /// See [Uri.queryParameters]
-  Map<String, dynamic> queryParameters;
 
   CancelToken? cancelToken;
 
   ProgressCallback? onReceiveProgress;
 
   ProgressCallback? onSendProgress;
-
-  int? connectTimeout;
 }
 
 /// The [_RequestConfig] class describes the http request information and configuration.
 class _RequestConfig {
   _RequestConfig({
-    this.method,
-    this.receiveTimeout,
-    this.sendTimeout,
+    int? receiveTimeout,
+    int? sendTimeout,
+    String? method,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
     String? contentType,
-    this.responseType,
-    this.validateStatus,
-    this.receiveDataWhenStatusError = true,
-    this.followRedirects = true,
-    this.maxRedirects = 5,
+    CollectionFormat? collectionFormat,
+    bool? followRedirects,
+    int? maxRedirects,
+    bool? receiveDataWhenStatusError,
+    ValidateStatus? validateStatus,
+    ResponseType? responseType,
     this.requestEncoder,
     this.responseDecoder,
   }) {
-    this.headers = caseInsensitiveKeyMap(value: headers);
-    this.extra = extra ?? {};
+    this.headers = caseInsensitiveKeyMap(headers);
     this.contentType = contentType;
+    this.method = method ?? 'GET';
+    this.sendTimeout = sendTimeout ?? 0;
+    this.receiveTimeout = receiveTimeout ?? 0;
+    this.collectionFormat = collectionFormat ?? CollectionFormat.csv;
+    this.extra = extra ?? {};
+    this.followRedirects = followRedirects ?? true;
+    this.maxRedirects = maxRedirects ?? 5;
+    this.receiveDataWhenStatusError = receiveDataWhenStatusError ?? true;
+    this.validateStatus = validateStatus ??
+        (int? status) {
+          return status != null && status >= 200 && status < 300;
+        };
+    this.responseType = responseType ?? ResponseType.json;
+    this.contentType = contentType ?? Headers.jsonContentType;
   }
 
   /// Http method.
-  String? method;
+  late String method;
 
   /// Http request headers. The keys of initial headers will be converted to lowercase,
   /// for example 'Content-Type' will be converted to 'content-type'.
   ///
   /// You should use lowercase as the key name when you need to set the request header.
-  Map<String, dynamic> headers;
+  late Map<String, dynamic> headers;
 
   /// Timeout in milliseconds for sending data.
   /// [Dio] will throw the [DioError] with [DioErrorType.SEND_TIMEOUT] type
   ///  when time out.
-  int? sendTimeout;
+  late int sendTimeout;
 
   ///  Timeout in milliseconds for receiving data.
   ///  [Dio] will throw the [DioError] with [DioErrorType.RECEIVE_TIMEOUT] type
   ///  when time out.
   ///
   /// [0] meanings no timeout limit.
-  int? receiveTimeout;
+  late int receiveTimeout;
 
   /// The request Content-Type. The default value is [ContentType.json].
   /// If you want to encode request body with 'application/x-www-form-urlencoded',
@@ -399,28 +557,28 @@ class _RequestConfig {
   ///
   /// If you want to receive the response data with  original bytes,
   /// that's to say the type of [Response.data] will be List<int>, use `bytes`
-  ResponseType? responseType;
+  late ResponseType responseType;
 
   /// `validateStatus` defines whether the request is successful for a given
   /// HTTP response status code. If `validateStatus` returns `true` ,
   /// the request will be perceived as successful; otherwise, considered as failed.
-  ValidateStatus? validateStatus;
+  late ValidateStatus validateStatus;
 
   /// Whether receiving response data when http status code is not successful.
-  bool? receiveDataWhenStatusError;
+  late bool receiveDataWhenStatusError;
 
   /// Custom field that you can retrieve it later in [Interceptor]、[Transformer] and the [Response] object.
-  Map<String, dynamic> extra;
+  late Map<String, dynamic> extra;
 
   /// see [HttpClientRequest.followRedirects]
-  bool? followRedirects;
+  late bool followRedirects;
 
   /// Set this property to the maximum number of redirects to follow
   /// when [followRedirects] is `true`. If this number is exceeded
   /// an error event will be added with a [RedirectException].
   ///
   /// The default value is 5.
-  int? maxRedirects;
+  late int maxRedirects;
 
   /// The default request encoder is utf8encoder, you can set custom
   /// encoder by this option.
@@ -429,4 +587,9 @@ class _RequestConfig {
   /// The default response decoder is utf8decoder, you can set custom
   /// decoder by this option, it will be used in [Transformer].
   ResponseDecoder? responseDecoder;
+
+  /// [collectionFormat] indicates the format of collection data in request
+  /// options which defined in [CollectionFormat] are `csv`, `ssv`, `tsv`, `pipes`, `multi`.
+  /// The default value is `csv`
+  late CollectionFormat collectionFormat;
 }
