+++ dio/pubspec.yaml
@@ -1,10 +1,10 @@
 name: dio
 description: A powerful Http client for Dart, which supports Interceptors, FormData, Request Cancellation, File Downloading, Timeout etc.
-version: 3.0.10
+version: 4.0.0-beta
 homepage: https://github.com/flutterchina/dio
 
 environment:
-  sdk: ">=2.12.0-0 <3.0.0"
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dependencies:
   http_parser: ^4.0.0
