+++ dio/test/_formdata
@@ -12,7 +12,7 @@ content-type: text/plain; charset=utf-8
 
 hellow world.
 ----dio-boundary-3788753558
-content-disposition: form-data; name="files[]"; filename="1.txt"
+content-disposition: form-data; name="files"; filename="1.txt"
 content-type: application/octet-stream
 
 你好世界，
@@ -20,7 +20,7 @@ content-type: application/octet-stream
 哈哈哈哈哈😆
 
 ----dio-boundary-3788753558
-content-disposition: form-data; name="files[]"; filename="2.txt"
+content-disposition: form-data; name="files"; filename="2.txt"
 content-type: application/octet-stream
 
 你好世界，
