+++ dio/test/basic_test.dart
@@ -61,8 +61,8 @@ void main() {
     assert(opt3.method == 'post');
     assert(opt3.receiveTimeout == 3000);
     assert(opt3.followRedirects == false);
-    assert(opt3.headers['b'] == '6');
-    assert(opt3.extra['b'] == '6');
+    assert(opt3.headers!['b'] == '6');
+    assert(opt3.extra!['b'] == '6');
     assert(opt3.contentType == 'text/html');
 
     var opt4 = RequestOptions(
@@ -142,7 +142,7 @@ void main() {
         throwsA(isTrue));
   });
 
-  test('#url encode ', () {
+  test('#url encode default ', () {
     var data = {
       'a': '你好',
       'b': [5, '6'],
@@ -155,7 +155,7 @@ void main() {
       }
     };
     var result =
-        'a=%E4%BD%A0%E5%A5%BD&b%5B%5D=5&b%5B%5D=6&c%5Bd%5D=8&c%5Be%5D%5Ba%5D=5&c%5Be%5D%5Bb%5D%5B%5D=66&c%5Be%5D%5Bb%5D%5B%5D=8';
+        'a=%E4%BD%A0%E5%A5%BD&b=5%2C6&c%5Bd%5D=8&c%5Be%5D%5Ba%5D=5&c%5Be%5D%5Bb%5D=66%2C8';
     expect(Transformer.urlEncodeMap(data), result);
   });
 }
