+++ example/options.dart
@@ -41,9 +41,8 @@ main() async {
   );
   print(response.data);
 
-  response = await dio.request(
-    "/",
-    options: RequestOptions(baseUrl: "https://baidu.com"),
+  response = await dio.fetch(
+    RequestOptions(baseUrl: "https://baidu.com/"),
   );
   print(response.data);
 }
