+++ plugins/cookie_manager/lib/src/cookie_mgr.dart
@@ -14,13 +14,7 @@ class CookieManager extends Interceptor {
   @override
   Future onRequest(RequestOptions options) async {
     var cookies = cookieJar.loadForRequest(options.uri);
-    cookies.removeWhere((cookie) {
-      if (cookie.expires != null) {
-        return cookie.expires.isBefore(DateTime.now());
-      }
-      return false;
-    });
-    String cookie = getCookies(cookies);
+    var cookie = getCookies(cookies);
     if (cookie.isNotEmpty) options.headers[HttpHeaders.cookieHeader] = cookie;
   }
 
@@ -30,19 +24,17 @@ class CookieManager extends Interceptor {
   @override
   Future onError(DioError err) async => _saveCookies(err.response);
 
-  _saveCookies(Response response) {
-    if (response != null && response.headers != null) {
-      List<String> cookies = response.headers[HttpHeaders.setCookieHeader];
+  void _saveCookies(Response? response) {
+      var cookies = response?.headers[HttpHeaders.setCookieHeader];
       if (cookies != null) {
         cookieJar.saveFromResponse(
-          response.request.uri,
+          response!.request.uri,
           cookies.map((str) => Cookie.fromSetCookieValue(str)).toList(),
         );
       }
-    }
   }
 
   static String getCookies(List<Cookie> cookies) {
-    return cookies.map((cookie) => "${cookie.name}=${cookie.value}").join('; ');
+    return cookies.map((cookie) => '${cookie.name}=${cookie.value}').join('; ');
   }
 }
