+++ plugins/cookie_manager/pubspec.yaml
@@ -1,20 +1,19 @@
 name: dio_cookie_manager
 description: A cookie manager for Dio, which supports persistent cookies in RAM and file.
-version: 1.0.0
+version: 2.0.0
 homepage:  https://github.com/flutterchina/dio/tree/master/plugins/cookie_manager
 author: wendux <824783146@qq.com>
 
 environment:
-  sdk: '>2.4.0 <3.0.0'
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dependencies:
-  cookie_jar: ^1.0.1
-#  dio: ^3.0.0
+  cookie_jar: ^2.0.0
+#  dio: ^4.0.0
   dio:
     path: ../../dio
 
 dev_dependencies:
-  test: ^1.5.1
-  test_coverage: ^0.2.0
-  pedantic: '1.1.0'
+  test: ^1.16.4
+  pedantic: ^1.10.0
