+++ plugins/http2_adapter/pubspec.yaml
@@ -1,19 +1,18 @@
 name: dio_http2_adapter
 description: A Dio HttpClientAdapter which support Http/2.0(Support connection reuse, header compression, server pushing is not supported yet.)
-version: 1.0.1
+version: 2.0.0
 homepage: https://github.com/flutterchina/dio/tree/master/plugins/http2_adapter
 
 environment:
-  sdk: '>2.4.0 <3.0.0'
+  sdk: '>=2.12.0-29.10.beta <3.0.0'
 
 dependencies:
-  http2: ^1.0.0
-  dio: ^3.0.0
-#  dio:
-#    path: ../../dio
+  http2: ^2.0.0
+#  dio: ^4.0.0
+  dio:
+    path: ../../dio
 
 dev_dependencies:
-  test: ^1.5.1
-  test_coverage: ^0.2.0
-  pedantic: '1.1.0'
+  test: ^1.16.4
+  pedantic: ^1.10.0
